+++ src/plugins/ordering/index.ts
@@ -1,2 +1,4 @@
 export * from './no/noOrdering.component';
 export * from './no/noOrdering.interface';
+export * from './single/singleOrdering.component';
+export * from './single/singleOrdering.interface';